improve arithmetics accuracy by (int -> float) fixed length

+++ Project_1/Models/Constraints/FixedLength.cs
@@ -3,12 +3,12 @@ using Project_1.Models.Shapes.Abstract;
 
 namespace Project_1.Models.Constraints
 {
-    public class FixedLength : IEdgeConstraint<int>
+    public class FixedLength : IEdgeConstraint<float>
     {
         public IEdge Edge { get; private set; }
-        public int Value { get; set; }
+        public float Value { get; set; }
 
-        public FixedLength(IEdge edge, int length)
+        public FixedLength(IEdge edge, float length)
         {
             Edge = edge;
             Value = length;
