+++ Project_1/Models/Repositories/Abstract/IConstraintRepositories.cs
@@ -5,7 +5,7 @@ namespace Project_1.Models.Repositories.Abstract
 {
     public interface IConstraintRepositories
     {
-        IEdgeConstraintRepository<FixedLength, int> FixedLengthRepository { get; }
+        IEdgeConstraintRepository<FixedLength, float> FixedLengthRepository { get; }
         IEdgeConstraintRepository<Perpendicular, IEdge> PerpendicularRepository { get; }
 
         void RemoveAllForEdge(IEdge edge);
